Moved OAS mesh generation to own files
