+++ openconcept/aerodynamics/openaerostruct/__init__.py
@@ -1,2 +1,3 @@
-from .drag_polar import VLMDragPolar, VLMDataGen, VLM, PlanformMesh
+from .mesh_gen import TrapezoidalPlanformMesh
+from .drag_polar import VLMDragPolar, VLMDataGen, VLM
 from .aerostructural import AerostructDragPolar, OASDataGen, Aerostruct, AerostructDragPolarExact
